trying to get all items with pagination v2

+++ src/main/kotlin/sidim/doma/undying/dto/PageDto.kt
@@ -2,7 +2,7 @@ package sidim.doma.undying.dto
 
 data class PageDto<T>(
     val content: List<T>,
-    val totalElements: Long,
+    val totalElements: Int,
     val totalPages: Int,
     val currentNumberPage: Int
 )
